Extract PreExecWorkingSet to separate file
- Created scratchpad/pre_exec_working_set.rs with PreExecWorkingSet and all impls
- Updated scratchpad/mod.rs to re-export PreExecWorkingSet
- Removed duplicate code from mod.rs
- Fields are pub(super) for sibling module access
- Code compiles with no warnings
